Rename main application PrimeMine->WinPrime
