+++ LICENSE
@@ -1,4 +1,4 @@
-PrimeMine - An easy-to-use Primecoin Client for Windows
+WinPrime - An easy-to-use Primecoin Node for Windows
 Copyright (C) 2018 Chapman Shoop
 
 This program is free software: you can redistribute it and/or modify
