+++ README.md
@@ -1,8 +1,8 @@
-# PrimeMine
+# WinPrime
 
 [![License: GPL v3](https://img.shields.io/badge/License-GPL%20v3-blue.svg)](https://www.gnu.org/licenses/gpl-3.0)
 
-An easy-to-use Primecoin Client (Network, Blockchain, Mine and Wallet) :)
+An easy-to-use Primecoin Node (Network, Blockchain, Mine and Wallet) :)
 
 Written in C# on Windows for Windows :D (Thanks for the Visual Studio!)
