+++ WinPrime/Properties/Resources.resx
