+++ WinPrime/WinPrime.csproj
@@ -6,8 +6,8 @@
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProjectGuid>{F420695D-C246-475E-A94E-E3E5A89565C9}</ProjectGuid>
     <OutputType>WinExe</OutputType>
-    <RootNamespace>PrimeMine</RootNamespace>
-    <AssemblyName>PrimeMine</AssemblyName>
+    <RootNamespace>WinPrime</RootNamespace>
+    <AssemblyName>WinPrime</AssemblyName>
     <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <ProjectTypeGuids>{60dc8134-eba5-43b8-bcc9-bb4bc16c2548};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
