+++ primecoin-windows.sln
@@ -3,11 +3,11 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.27428.2037
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PrimeMine", "PrimeMine\PrimeMine.csproj", "{F420695D-C246-475E-A94E-E3E5A89565C9}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PrimeNetwork", "PrimeNetwork\PrimeNetwork.csproj", "{3905A3EA-5CC9-4FC8-BA8C-DF4B6F5BAA3E}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PrimeNetworkTest", "PrimeNetworkTest\PrimeNetworkTest.csproj", "{CDB9167E-8BCC-4B74-B837-3B8D9F9F5343}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WinPrime", "WinPrime\WinPrime.csproj", "{F420695D-C246-475E-A94E-E3E5A89565C9}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,10 +15,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F420695D-C246-475E-A94E-E3E5A89565C9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F420695D-C246-475E-A94E-E3E5A89565C9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F420695D-C246-475E-A94E-E3E5A89565C9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F420695D-C246-475E-A94E-E3E5A89565C9}.Release|Any CPU.Build.0 = Release|Any CPU
 		{3905A3EA-5CC9-4FC8-BA8C-DF4B6F5BAA3E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{3905A3EA-5CC9-4FC8-BA8C-DF4B6F5BAA3E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3905A3EA-5CC9-4FC8-BA8C-DF4B6F5BAA3E}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -27,6 +23,10 @@ Global
 		{CDB9167E-8BCC-4B74-B837-3B8D9F9F5343}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CDB9167E-8BCC-4B74-B837-3B8D9F9F5343}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CDB9167E-8BCC-4B74-B837-3B8D9F9F5343}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F420695D-C246-475E-A94E-E3E5A89565C9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F420695D-C246-475E-A94E-E3E5A89565C9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F420695D-C246-475E-A94E-E3E5A89565C9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F420695D-C246-475E-A94E-E3E5A89565C9}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
